Sandboxer: Replace features with bin parameter
