+++ .gitignore
@@ -1,5 +1,5 @@
 .idea
-bin
+/bin
 .vscode
 target
 **/src/api/*
