+++ vmm/sandbox/Cargo.toml
@@ -7,12 +7,6 @@ edition = "2021"
 [profile.release]
 panic = 'abort'
 
-[features]
-default = []
-qemu = ["qapi/qmp", "qapi/async-tokio-all", "unshare", "os_pipe"]
-cloud_hypervisor = ["api_client"]
-stratovirt = ["qapi/qmp", "qapi/async-tokio-all", "unshare", "os_pipe"]
-
 [dependencies]
 tokio = { version = "1.19.2", features = ["full"] }
 containerd-sandbox = {git="https://github.com/kuasar-io/rust-extensions.git"}
@@ -36,15 +30,28 @@ prost-types = "0.10.1"
 time = "0.3.5"
 log = { version = "0.4.17", features = ["std"] }
 uuid = { version = "1.1.2", features = ["v4"] }
-unshare = { version = "0.7.0", optional = true }
-os_pipe = { version = "0.9.2", optional = true }
-qapi = { version = "0.8.0", features = ["qmp", "async-tokio-all"], optional = true }
+unshare = { version = "0.7.0" }
+os_pipe = { version = "0.9.2" }
+qapi = { version = "0.8.0", features = ["qmp", "async-tokio-all"] }
 qapi-spec = {version = "0.3.1"}
 sandbox-derive = { path = "derive" }
-api_client = { git = "https://github.com/cloud-hypervisor/cloud-hypervisor.git", optional = true }
-rtnetlink = "0.12"
-netlink-packet-route = "0.15"
-netlink-packet-core = "0.5.0"
+api_client = { git = "https://github.com/cloud-hypervisor/cloud-hypervisor.git" }
+rtnetlink = "0.13.1"
+netlink-packet-route = "0.17.0"
+netlink-packet-core = "0.7.0"
 ttrpc = { version = "0.7", features = ["async"] }
 protobuf = "3.2"
 cgroups-rs = "0.3.2"
+proc-macro2 = "1.0.66"
+
+[[bin]]
+name = "qemu" 
+path = "src/bin/qemu/main.rs"
+
+[[bin]]
+name = "cloud_hypervisor" 
+path = "src/bin/cloud_hypervisor/main.rs"
+
+[[bin]]
+name = "stratovirt" 
+path = "src/bin/stratovirt/main.rs"
